draft stash: dialog-safe clear sequence + one user-input write path (PRDCT-1569 fix loop)
Independent verification of 5802e83 proved two behavior defects on the real
CLI; both fixed here, with the record corrected.

F1 — the clear sequence moved the /model effort control. The degraded-
confidence clear emitted cpAll+1024 right-arrows before its backspaces.
Right-arrow is inert in the input box but adjusts the effort control in the
CLI's /model dialog, and a degraded shadow is exactly the state in which that
dialog may be open (user typed /model + Enter, then Down to browse); the
dispatcher's next write is the submitting CR, which persists the change.
The clear now uses forward-delete (\x1b[3~) for everything after the cursor
and backspace (\x7f) for the full length — no right-arrows anywhere, on the
confident path either. Both keys were probed on the real CLI first: 64
presses each leave /model untouched (Right x3 moved it xHigh->Low as the
positive control), Delete forward-deletes in the input ("abc", Left, Left,
Delete -> "ac"), and 64 Deletes on an empty input are a no-op with the CLI
alive. Ctrl+E and End were also inert in /model but are line-end keys,
useless for multi-line drafts; forward-delete is count-based.

F2 — the file palette bypassed the shadow. FilePalette -> shell.ts insertPath
wrote through writeSession directly, so a palette-inserted path followed by a
sibling send co-submitted exactly per the ticket. There is now ONE user-input
write helper, lib/user-input.ts writeUserInput(), and every renderer writer
of user input goes through it: use-terminal (onData, custom key bindings,
drag-drop), use-toolbar-terminal (onData), and shell.ts insertPath. The
remaining direct writeSession callers are app-originated, not user input,
and stay as they are: the dispatcher's own injection writes (accounted for
via begin/endInjection) and the server-button command/Ctrl+C writes into
plain terminal sessions (AppShell, TerminalHeader, use-server-button), which
are never send targets. Remote tabs write via sshShellWrite and are never
send targets either.

Record corrections to 5802e83's message: the degradation is NOT visible to
the user — draftHandling surfaces only in the tool result, nothing in the app
UI shows it (the injected-from badge marks every send alike); and "boundary
presses are no-ops" held for the input box only, not for dialogs — the clear
sequence is now built solely from keys verified for both.

Verified live (playwright-core driving the built app, real claude tabs, real
MCP HTTP sends): idle draft, two concurrent sends mid-turn, palette-inserted
path (restored, not co-submitted), and /model open with a degraded shadow
(effort xhigh before and after the send, same model, best-effort reported);
46-check standalone node probe of the pure module passes.
